Cleanup JVP optimization experiments, finalize MeanFlow settings
Changes:
- Delete experimental cruft: fast_jvp_norm.py, benchmark scripts for
  discarded optimization approaches (split_jvp, microbatch, forward_ad)
- Fix SDPA attention to use manual attention on MPS (SDPA doesn't
  support forward-mode AD on MPS)
- Fix dtype mismatch in selective JVP mode under bf16 autocast
- Add CUDA graph options (full_batch_jvp, use_cuda_graph) to meanflow
  config with documentation
- Update CLAUDE.md and README.md with final MeanFlow performance
  recommendations (~19% overhead at 25% ratio with CUDA graph + BF16)
- Add bf16_training.md design doc
- Add .codex_tmp/ to gitignore

Performance summary (A100, UNet):
- MeanFlow 100% (default selective): +237% overhead
- MeanFlow 100% (CUDA graph + BF16): +72% overhead
- MeanFlow 25% (CUDA graph + BF16): ~19% overhead (paper claims 16%)

+++ losses/meanflow_loss.py
@@ -554,12 +554,12 @@ class MeanFlowLoss(nn.Module):
             if fm_mask.any():
                 unified_fm = unified_t[fm_mask]
                 u_pred_fm = self._model(unified_fm)
-                u_pred[fm_mask] = u_pred_fm
+                u_pred[fm_mask] = u_pred_fm.to(u_pred.dtype)
 
             # Use JVP primal for MeanFlow samples (if any exist)
             if mf_pred_info is not None:
                 mf_mask_from_jvp, u_pred_mf = mf_pred_info
-                u_pred[mf_mask_from_jvp] = u_pred_mf
+                u_pred[mf_mask_from_jvp] = u_pred_mf.to(u_pred.dtype)
 
         # Compute squared error per sample (sum over spatial dims, not mean)
         # This matches official: loss.sum(dim=(1, 2, 3))
